# crm/docs/openapi-mvp.yaml
openapi: 3.1.0
info:
  title: CRM Automotivo MVP API
  version: 0.1.0
  description: Contrato inicial da API Fastify do CRM automotivo MVP.
servers:
  - url: http://localhost:3333
    description: Desenvolvimento local
security:
  - bearerAuth: []
...
paths:
...
  /sales/{id}:
    get: { tags: [Sales], summary: Detalha venda., parameters: [ { $ref: "#/components/parameters/Id" } ], responses: { "200": { $ref: "#/components/responses/Ok" } } }
    patch: { tags: [Sales], summary: Atualiza venda., parameters: [ { $ref: "#/components/parameters/Id" } ], responses: { "200": { $ref: "#/components/responses/Ok" } } }
  /sales/{id}/status:
    post: { tags: [Sales], summary: Altera status/aprova/fecha venda., parameters: [ { $ref: "#/components/parameters/Id" } ], responses: { "200": { $ref: "#/components/responses/Ok" } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Ok:
      description: Resposta bem-sucedida.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
...